KETO  - Ecommerce Product
Keto is an e-commerce platform designed for showcasing and selling electronic devices like smartphones, tablets, and accessories. It features a sleek user interface, product filtering, and secure cart and checkout functionalities. The platform offers real-time product updates and an admin dashboard for inventory management. Keto ensures a smooth shopping experience with a focus on modern electronics.